requirements.txt
paho-mqtt==2.1.0
redis==7.1.0
